pyproject.toml
[project]
name = "Telegram-Linux-Manager"
version = "0.3.0"
description = "A simple server information manager about your linux server."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.5",
    "requests>=2.32.5",
]
